# backend/api/routes/disputeRoutes.js
import express from 'express';
import disputeController from '../controllers/disputeController.js';
import { cacheResponse, invalidateOn, TTL } from '../middleware/cache.js';
import authMiddleware from '../middleware/auth.js';
import { requireMfa } from '../middleware/mfaAuth.js';
import { handleUploadError } from '../middleware/fileUpload.js';
import {
  validate,
  disputeListQueryRules,
  disputeEscrowIdParamRules,
} from '../middleware/validation.js';

const router = express.Router();
router.use(authMiddleware);

// ── Create ────────────────────────────────────────────────────────────────────

/** POST /api/disputes — raise a dispute; requires categoryId, auto-assigns arbiter pool */
router.post('/', disputeController.createDispute);

// ── List / Get ────────────────────────────────────────────────────────────────

router.get(
  '/',
  validate(disputeListQueryRules),
  cacheResponse({ ttl: TTL.LIST, tags: ['disputes'] }),
  disputeController.listDisputes,
);

router.get(
  '/history',
  cacheResponse({ ttl: TTL.LIST, tags: ['disputes', 'disputes:history'] }),
  disputeController.getResolutionHistory,
);

router.get(
  '/:escrowId',
  validate(disputeEscrowIdParamRules),
  cacheResponse({
    ttl: TTL.DETAIL,
    tags: (req) => ['disputes', `dispute:${req.params.escrowId}`],
  }),
  disputeController.getDispute,
);

// ── Timeline ──────────────────────────────────────────────────────────────────

router.get(
  '/:id/timeline',
  cacheResponse({
    ttl: TTL.DETAIL,
    tags: (req) => [`dispute:${req.params.id}`],
  }),
  disputeController.getTimeline,
);

// ── Evidence ──────────────────────────────────────────────────────────────────

router.post(
  '/:id/evidence',
  invalidateOn({ tags: (req) => [`dispute:${req.params.id}`, 'disputes'] }),
  disputeController.uploadEvidence,
  disputeController.postEvidence,
  handleUploadError,
);

router.get(
  '/:id/evidence',
  cacheResponse({
    ttl: TTL.DETAIL,
    tags: (req) => [`dispute:${req.params.id}`],
  }),
  disputeController.listEvidence,
);

// ── Automated Resolution ──────────────────────────────────────────────────────

router.post(
  '/:id/resolve/auto',
  requireMfa,
  invalidateOn({
    tags: (req) => [`dispute:${req.params.id}`, `escrow:${req.params.id}`, 'disputes', 'escrows'],
  }),
  disputeController.autoResolve,
);

router.get(
  '/:id/resolve/recommendation',
  cacheResponse({
    ttl: TTL.DETAIL,
    tags: (req) => [`dispute:${req.params.id}`],
  }),
  disputeController.getRecommendation,
);

// ── Appeals ───────────────────────────────────────────────────────────────────

router.post(
  '/:id/appeals',
  invalidateOn({ tags: (req) => [`dispute:${req.params.id}`, 'disputes'] }),
  disputeController.postAppeal,
);

router.patch(
  '/appeals/:appealId',
  requireMfa,
  invalidateOn({ tags: ['disputes'] }),
  disputeController.patchAppeal,
);

export default router;