# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar Trust Escrow API
  version: 0.1.0
  description: |
    REST API for the Stellar Trust Escrow platform — a trust-minimised escrow
    service built on the Stellar blockchain.

    ## Authentication
    Most endpoints require a Bearer JWT token in the `Authorization` header.
    Admin endpoints additionally require the `x-admin-api-key` header.

    ## Pagination
    All collection endpoints use **cursor-based pagination** and return:
    ```
    {
      "data": [...],
      "next_cursor": "<opaque-base64url-token>",
      "has_more": true
    }
    ```
    Pass `?cursor=<next_cursor>` to fetch the next page. The cursor encodes the
    sort key and direction and is stable across concurrent inserts.

    ## Idempotency Keys
    All `POST` and `PATCH` endpoints accept an optional `Idempotency-Key` header.
    Supplying this header makes the call safe to retry:
    - The response is cached in Redis for **24 hours** under `(tenantId, method, path, key)`.
    - Duplicate requests within the TTL return the cached response immediately
      (status `X-Idempotency-Replayed: true`).
    - If a duplicate arrives while the first is still in-flight, a **409 Conflict** is returned.
    - Use any unique string (UUID recommended): `Idempotency-Key: <uuid-v4>`

    ## Environments
    - **Testnet**: `https://api.test.stellar-trust-escrow.example.com`
    - **Mainnet**: `https://api.stellar-trust-escrow.example.com`
  contact:
...
  license:
    name: AGPL-3.0
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login with email and password
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: MyStr0ng!Pass
      responses:
        '200':
          description: JWT tokens returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/users/{address}:
    get:
      tags: [Users]
      summary: Get user profile (reputation, escrow history, stats)
      operationId: getUserProfile
      security:
        - BearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          description: Stellar public key (G...)
          schema:
            type: string
            example: GDKHX...QZ4
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/users/{address}/stats:
    get:
      tags: [Users]
      summary: Aggregated user statistics
      operationId: getUserStats
      security:
        - BearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/disputes:
    get:
      tags: [Disputes]
      summary: List disputes with pagination
      operationId: listDisputes
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated dispute list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDisputes'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/disputes/{escrowId}:
    get:
      tags: [Disputes]
      summary: Get dispute details for a specific escrow
      operationId: getDispute
      security:
        - BearerAuth: []
      parameters:
        - name: escrowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dispute detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/stats:
    get:
      tags: [Admin]
      summary: Platform-wide statistics
      operationId: getAdminStats
      security:
        - AdminAuth: []
      responses:
        '200':
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List all users with pagination and search
      operationId: adminListUsers
      security:
        - AdminAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAdminUsers'
  /api/admin/users/{address}:
    get:
      tags: [Admin]
      summary: Get detailed user profile
      operationId: adminGetUser
      security:
        - AdminAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/users/{address}/suspend:
    post:
      tags: [Admin]
      summary: Suspend a user
      operationId: suspendUser
      security:
        - AdminAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  example: Terms of service violation
      responses:
        '200':
          description: User suspended
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/users/{address}/ban:
    post:
      tags: [Admin]
      summary: Permanently ban a user
      operationId: banUser
      security:
        - AdminAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: User banned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/disputes:
    get:
      tags: [Admin]
      summary: List all disputes with pagination
      operationId: adminListDisputes
      security:
        - AdminAuth: []
      parameters:
        - name: resolved
          in: query
          schema:
            type: boolean
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated dispute list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDisputes'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/disputes/{id}/resolve:
    post:
      tags: [Admin]
      summary: Resolve an open dispute
      operationId: resolveDispute
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientAmount, freelancerAmount]
              properties:
                clientAmount:
                  type: string
                  description: XLM amount to release to client
                  example: '250'
                freelancerAmount:
                  type: string
                  description: XLM amount to release to freelancer
                  example: '750'
                notes:
                  type: string
                  description: Internal resolution notes
      responses:
        '200':
          description: Dispute resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  txHash:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/settings:
    get:
      tags: [Admin]
      summary: Read current platform settings
      operationId: getSettings
      security:
        - AdminAuth: []
      responses:
        '200':
          description: Platform settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [Admin]
      summary: Update platform settings
      operationId: updateSettings
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                platformFeePercent:
                  type: number
                  minimum: 0
                  maximum: 100
                  example: 2.5
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/audit-logs:
    get:
      tags: [Admin]
      summary: Paginated admin action audit log
      operationId: getAdminAuditLogs
      security:
        - AdminAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLogs'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login
    AdminAuth:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Admin API key for protected admin endpoints
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number. Values below 1 are normalised to 1.
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      description: Page size. Defaults to 20 and is capped at 100.
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong
        errorId:
          type: string
          description: Sentry event ID for support correlation
    PaginatedResponse:
      type: object
      required: [data, page, limit, total, totalPages, hasNextPage, hasPreviousPage]
      properties:
        data:
          type: array
          items: {}
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3
        hasNextPage:
          type: boolean
          example: true
        hasPreviousPage:
          type: boolean
          example: false
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          example: 3600
        user:
          type: object
          properties:
            email:
              type: string
            stellarAddress:
              type: string
...
    UserProfile:
      type: object
      properties:
        address:
          type: string
        email:
          type: string
          nullable: true
        kycStatus:
          type: string
          enum: [Pending, Init, Processing, Approved, Declined]
        reputation:
          type: object
          properties:
            score:
              type: number
            totalReviews:
              type: integer
            completionRate:
              type: number
        escrowStats:
          type: object
          properties:
            asClient:
              type: integer
            asFreelancer:
              type: integer
            completed:
              type: integer
            disputed:
              type: integer
        createdAt:
          type: string
          format: date-time
    UserStats:
      type: object
      properties:
        address:
          type: string
        totalVolume:
          type: string
          description: Total XLM volume
        completionRate:
          type: number
          example: 0.92
        avgMilestoneTimeDays:
          type: number
        totalEscrows:
          type: integer
        completedEscrows:
          type: integer
        disputedEscrows:
          type: integer
...
    Dispute:
      type: object
      properties:
        id:
          type: string
        escrowId:
          type: string
        raisedBy:
          type: string
        reason:
          type: string
        status:
          type: string
          enum: [Open, Resolved]
        resolution:
          type: object
          nullable: true
          properties:
            clientAmount:
              type: string
            freelancerAmount:
              type: string
            notes:
              type: string
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
    PaginatedDisputes:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Dispute'
...
    AdminStats:
      type: object
      properties:
        totalEscrows:
          type: integer
        totalUsers:
          type: integer
        totalDisputes:
          type: integer
        openDisputes:
          type: integer
        totalVolumeXlm:
          type: string
        platformRevenueXlm:
          type: string
    AdminUserDetail:
      allOf:
        - $ref: '#/components/schemas/UserProfile'
        - type: object
          properties:
            banStatus:
              type: string
              enum: [active, suspended, banned]
            banReason:
              type: string
              nullable: true
            bannedAt:
              type: string
              format: date-time
              nullable: true
    PaginatedAdminUsers:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/AdminUserDetail'
    PlatformSettings:
      type: object
      properties:
        platformFeePercent:
          type: number
          example: 2.5
        minEscrowAmountXlm:
          type: string
        maxEscrowAmountXlm:
          type: string
        escrowLockPeriodDays:
          type: integer
        updatedAt:
          type: string
          format: date-time
    PaginatedAuditLogs:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/AuditLogEntry'
    AuditLogEntry:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
        action:
          type: string
        actor:
          type: string
        resourceId:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...